Struct sval::stream::OwnedStream

pub struct OwnedStream<S>(_);

An owned stream wrapper.

OwnedStream is an ergonomic wrapper over a raw Stream that makes it easier to stream complex types.

Methods

impl<S> OwnedStream<S> where
S: Stream,

pub fn stream(stream: S, value: impl Value) -> Result<S, Error>

Stream a value.

pub fn new(stream: S) -> Self

Begin an owned stream.

pub fn into_inner(self) -> S

Unwrap the inner stream.

pub fn borrow_mut(&mut self) -> RefMutStream

Get a reference to the stream that can be used by a value.

pub fn any(&mut self, v: impl Value) -> Result

Stream a value.

pub fn fmt(&mut self, f: Arguments) -> Result

Stream a format.

pub fn i64(&mut self, v: i64) -> Result

Stream a signed integer.

pub fn u64(&mut self, v: u64) -> Result

Stream an unsigned integer.

pub fn i128(&mut self, v: i128) -> Result

Stream a 128-bit signed integer.

pub fn u128(&mut self, v: u128) -> Result

Stream a 128-bit unsigned integer.

pub fn f64(&mut self, v: f64) -> Result

Stream a floating point value.

pub fn bool(&mut self, v: bool) -> Result

Stream a boolean.

pub fn char(&mut self, v: char) -> Result

Stream a unicode character.

pub fn str(&mut self, v: &str) -> Result

Stream a UTF8 string.

pub fn none(&mut self) -> Result

Stream an empty value.

pub fn map_begin(&mut self, len: Option<usize>) -> Result

Begin a map.

pub fn map_key(&mut self, k: impl Value) -> Result

Stream a map key.

pub fn map_value(&mut self, v: impl Value) -> Result

Stream a map value.

pub fn map_end(&mut self) -> Result

End a map.

pub fn seq_begin(&mut self, len: Option<usize>) -> Result

Begin a sequence.

pub fn seq_elem(&mut self, v: impl Value) -> Result

Stream a sequence element.

pub fn seq_end(&mut self) -> Result

End a sequence.

impl<S> OwnedStream<S> where
S: Stream,

pub fn map_key_begin(&mut self) -> Result<&mut Self, Error>

Begin a map key.

pub fn map_value_begin(&mut self) -> Result<&mut Self, Error>

Begin a map value.

pub fn seq_elem_begin(&mut self) -> Result<&mut Self, Error>

Begin a sequence element.