Struct sval::value::OwnedValue

pub struct OwnedValue(_);

An owned, immutable value.

Owned values are safe to share and are cheap to clone.

Add the std feature to your Cargo.toml to enable this type:

[dependencies.sval]
features = ["std"]

Methods

impl OwnedValue

pub fn collect(v: impl Value) -> Self

Get an owned value from an arbitrary Value.

The given value doesn't need to be Send + Sync + 'static.

Some primitive types can be converted into an OwnedValue for free. These types have a corresponding From implementation.

The structure of the given value will be streamed into a sharable representation. That means this method is more expensive for more complex values.

Prefer the From impls and from_shared method where possible.

pub fn from_shared(v: impl Into<Arc<dyn Value + Send + Sync>>) -> Self

Get an owned value from an already shared Value.

The given value must be Send + Sync + 'static.