Struct sval::value::Stream

pub struct Stream<'a>(_);

A borrowed stream wrapper.

This is the result of calling OwnedStream.borrow_mut.

Implementations

impl<'a> RefMutStream<'a>

pub fn any(&mut self, v: impl Value) -> Result

Stream a value.

pub fn debug(&mut self, v: impl Debug) -> Result

Stream a debuggable type.

pub fn display(&mut self, v: impl Display) -> Result

Stream a displayable type.

pub fn error(&mut self, v: &(dyn Error + 'static)) -> Result

Stream an error.

This method is only available when the std feature is enabled.

Examples

Errors that don't satisfy the trait bounds needed by this method can go through Source:

impl Value for MyError {
    fn stream(&self, stream: &mut value::Stream) -> value::Result {
        use sval::stream::Source;

        stream.any(Source::new(&self.error))
    }
}

pub fn i64(&mut self, v: i64) -> Result

Stream a signed integer.

pub fn u64(&mut self, v: u64) -> Result

Stream an unsigned integer.

pub fn i128(&mut self, v: i128) -> Result

Stream a 128-bit signed integer.

pub fn u128(&mut self, v: u128) -> Result

Stream a 128-bit unsigned integer.

pub fn f64(&mut self, v: f64) -> Result

Stream a floating point value.

pub fn bool(&mut self, v: bool) -> Result

Stream a boolean.

pub fn char(&mut self, v: char) -> Result

Stream a unicode character.

pub fn str(&mut self, v: &str) -> Result

Stream a UTF8 string.

pub fn none(&mut self) -> Result

Stream an empty value.

pub fn map_begin(&mut self, len: Option<usize>) -> Result

Begin a map.

pub fn map_key(&mut self, k: impl Value) -> Result

Stream a map key.

pub fn map_value(&mut self, v: impl Value) -> Result

Stream a map value.

pub fn map_end(&mut self) -> Result

End a map.

pub fn seq_begin(&mut self, len: Option<usize>) -> Result

Begin a sequence.

pub fn seq_elem(&mut self, v: impl Value) -> Result

Stream a sequence element.

pub fn seq_end(&mut self) -> Result

End a sequence.

impl<'a> RefMutStream<'a>

pub fn map_key_begin(&mut self) -> Result<&mut Self, Error>

Begin a map key.

pub fn map_value_begin(&mut self) -> Result<&mut Self, Error>

Begin a map value.

pub fn seq_elem_begin(&mut self) -> Result<&mut Self, Error>

Begin a sequence element.